Add pnm.list and mapping.csv
pnm.list contains a list of arguments given to BuildRequires in SFE specs
that use package name macros, with the leading "pnm_buildrequires_" removed.
It was produced by get-pn-symbols.pl.

mapping.csv shows what package the names in pnm.list are mapped to on S11.3,
OIhipster, and OmniOS.  It was produced by print-csv-mapping.pl, using the
results of running pkgnmmap.pl on each distribution.

(The input files to print-csv-mapping.pl were modified by removing
database/mysql-51/library from each, to keep the number of pn symbols and the
number of package names that those symbols map into identical.  For an unknown
reason, pkgnmmap.pl - as well as spectool, which it is based on - return two
packages when given "BuildRequires: %{pnm_buildrequires_mysql_default}".)